Updated Fuction for viewing data and also return page after submitting website

+++ src/main/java/com/example/webcrawler/repositories/WebsiteRepository.java
@@ -4,6 +4,6 @@ import com.example.webcrawler.controller.model.Website;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface WebsiteRepository extends JpaRepository<Website, Integer> {
+public interface WebsiteRepository extends JpaRepository<Website, Long> {
 
 }
